requirements.txt
asgiref==3.2.7
Babel==2.8.0
dj-database-url==0.5.0
Django==3.0.7
django-filter==2.3.0
django-heroku==0.3.1
django-phonenumber-field==4.0.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
phonenumbers==8.12.5
Pillow==7.1.2
psycopg2==2.8.5
Pygments==2.6.1
PyJWT==1.7.1
python-decouple==3.3
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.1.0
